[refactor] separate folders in interfaces folder
