+++ Backend/Interfaces/Services/IAdminService.cs
@@ -3,7 +3,7 @@ using Backend.Common;
 using Backend.Common.Params;
 using Backend.DTOs;
 
-namespace Backend.Interfaces;
+namespace Backend.Interfaces.Services;
 
 public interface IAdminService
 {
