+++ Backend/Interfaces/Services/IUserService.cs
@@ -2,7 +2,7 @@ using Backend.Common;
 using Backend.Common.Params;
 using Backend.DTOs;
 
-namespace Backend.Interfaces;
+namespace Backend.Interfaces.Services;
 
 public interface IUserService
 {
