+++ Backend/Repository/UserRepository.cs
@@ -2,7 +2,7 @@ using Backend.Common;
 using Backend.Common.Params;
 using Backend.DTOs;
 using Backend.Entities;
-using Backend.Interfaces;
+using Backend.Interfaces.Repository;
 using Microsoft.EntityFrameworkCore;
 
 namespace Backend.Data;
